feat: put strategy content inside crab page
